requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
Flask-RESTful==0.3.9
fonttools==4.32.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
Werkzeug==2.1.1
